+++ examples/bins/plot_ah_coordinate_velocity.py
@@ -105,5 +105,10 @@ given apparent horizon as a function of time."""
         logger.debug("Plotted")
 
         logger.debug("Saving")
-        save_from_dir_filename_ext(args.outdir, figname, args.fig_extension)
+        save_from_dir_filename_ext(
+            args.outdir,
+            figname,
+            args.fig_extension,
+            tikz_clean_figure=args.tikz_clean_figure,
+        )
         logger.debug("DONE")