+++ examples/bins/plot_ah_found.py
@@ -99,5 +99,10 @@ if __name__ == "__main__":
         logger.debug("Plotted")
 
         logger.debug("Saving")
-        save_from_dir_filename_ext(args.outdir, figname, args.fig_extension)
+        save_from_dir_filename_ext(
+            args.outdir,
+            figname,
+            args.fig_extension,
+            tikz_clean_figure=args.tikz_clean_figure,
+        )
         logger.debug("DONE")