+++ examples/bins/plot_ah_radius.py
@@ -114,5 +114,10 @@ all covered by the given resolution."""
         logger.debug("Plotted")
 
         logger.debug("Saving")
-        save_from_dir_filename_ext(args.outdir, figname, args.fig_extension)
+        save_from_dir_filename_ext(
+            args.outdir,
+            figname,
+            args.fig_extension,
+            tikz_clean_figure=args.tikz_clean_figure,
+        )
         logger.debug("DONE")