+++ examples/bins/plot_constraints.py
@@ -104,7 +104,10 @@ if __name__ == "__main__":
 
             logger.debug("Saving")
             save_from_dir_filename_ext(
-                args.outdir, figname, args.fig_extension
+                args.outdir,
+                figname,
+                args.fig_extension,
+                tikz_clean_figure=args.tikz_clean_figure,
             )
             logger.debug("Saved")