+++ examples/bins/plot_em_energy.py
@@ -97,5 +97,10 @@ luminosity and cumulative energy as a function of time as computed from Phi2.
         logger.debug("Plotted")
 
         logger.debug("Saving")
-        save_from_dir_filename_ext(args.outdir, figname, args.fig_extension)
+        save_from_dir_filename_ext(
+            args.outdir,
+            figname,
+            args.fig_extension,
+            tikz_clean_figure=args.tikz_clean_figure,
+        )
         logger.debug("DONE")