+++ kuibit/visualize_matplotlib.py
@@ -370,6 +370,7 @@ def save(
     outputpath,
     figure=None,
     axis=None,
+    tikz_clean_figure=False,
     **kwargs,
 ):
     """Save figure to the given location.
@@ -378,7 +379,9 @@ def save(
     ``tikzplotlib``.
 
     Unknown arguments are passed to the ``matplotlib.savefig`` or
-    ``tikzplotlib.save`` (depending on the extension).
+    ``tikzplotlib.save`` (depending on the extension). In this second case, if
+    ``tikz_clean_figure = True``, unknown arguments are first passed to
+    ``tikzplotlib.clean_figure``.
 
     :param outputpath:  Output path with or without extension. If the
                         extension is ``.tikz``, the file is saved with
@@ -388,12 +391,44 @@ def save(
                    use the current figure.
     :type figure: ``matplotlib.pyplot.figure``
 
+    :param tikz_clean_figure: If ``tikzplotlib`` is begin used, reduce the size
+                              of the output ``tikz`` file. When this is set to
+                              True, unknown arguments are first passed to
+                              ``tikzplotlib.clean_figure``, then to
+                              ``tikzplotlib.save``. ``tikzplotlib.clean_figure``
+                              will change the given figure.
+    :type tikz_clean_figure: bool
+
     :param axis: If passed, plot on this axis. If not passed (or if None), use
                  the current axis.
     :type axis: ``matplotlib.pyplot.axis``
 
     """
     if os.path.splitext(outputpath)[-1] == ".tikz":
+
+        # If clean_figure is True, we extract from kwargs those argument
+        # that tikzplotlib.clean_figure would take. For this, we need to
+        # know what argument that function takes.
+        #
+        # Form https://stackoverflow.com/a/40363565
+        if tikz_clean_figure:
+            args_names = tikzplotlib.clean_figure.__code__.co_varnames[
+                : tikzplotlib.clean_figure.__code__.co_argcount
+            ]
+
+            kwargs_clean_figure = {}
+
+            # We split kwargs in those that are supposed to be passed to
+            # clean_figure and those that have to be passed to save. For this,
+            # we iterate over the arguments taken by clean_figure, if they are
+            # in kwargs, then we move them to a new dictionary
+            for arg in args_names:
+                if arg in kwargs:
+                    kwargs_clean_figure.update({arg: kwargs[arg]})
+                    del kwargs[arg]
+
+            tikzplotlib.clean_figure(fig=figure, **kwargs_clean_figure)
+
         tikzplotlib.save(outputpath, **kwargs)
     else:
         figure.savefig(outputpath, **kwargs)
@@ -406,6 +441,7 @@ def save_from_dir_filename_ext(
     file_ext,
     figure=None,
     axis=None,
+    tikz_clean_figure=False,
     **kwargs,
 ):
     """Save figure to a location defined by a folder, a name, and an extension.
@@ -437,6 +473,7 @@ def save_from_dir_filename_ext(
         outputpath,
         figure=figure,
         axis=axis,
+        tikz_clean_figure=tikz_clean_figure,
         **kwargs,
     )